+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - added shift-click support for selecting task ranges in multiselect mode
 - correctly showing uncut list of uncommited files in the worktree status
 - added task actions based on the task state for easier workflow
+- added task settings page with smart task state and auto-generate task name configuration
 
 ## [0.46.0]